Semantic?
This seems like a more correct way to handle errors. It seems like the BadEncoding error was being misused, since it was not describing an error related to encoding. I put in expect instead bc it was simple and seems easy to debug which is appealing bc I don;t have a good understanding yet.